refactor(#742): relocate tail dispatch to call-tail-dispatch.ts — decomposition DONE
Wave B chunk 5, step 2 (final). Move compileTailDispatch (IIFE, super,
element-access, call-of-call, conditional callee, graceful fallback) out of
calls.ts into a new sibling module src/codegen/expressions/call-tail-dispatch.ts.
12 calls.ts internals imported (6 newly exported); calls.ts imports
compileTailDispatch back (lazy cycle).

calls.ts: 9,534 -> 7,753 LOC. compileCallExpression: ~3,430 -> ~1,685 LOC. New
call-tail-dispatch.ts: 1,872 LOC.

Completes the compileCallExpression decomposition: 13,371 -> ~1,685 LOC (-87%)
across 5 byte-identical slices; calls.ts 19,435 -> 7,753. compileCallExpression
is now a lean dispatch skeleton delegating to per-shape sibling modules. #742
status: done.

Pure relocation — emitted Wasm byte-identical (prove-emit-identity IDENTICAL,
39/39). tsc 0. tests/issue-742.test.ts 16/16 (added IIFE + super cases).
loc-budget + coercion-sites (2 relocated sites, net-zero) pass with allowances;
oracle-ratchet net-zero.

✓
